+++ crates/teloxide-core/src/adaptors/erased.rs
@@ -941,7 +941,7 @@ trait ErasableRequester<'a> {
 
     fn send_game(
         &self,
-        chat_id: u32,
+        chat_id: ChatId,
         game_short_name: String,
     ) -> ErasedRequest<'a, SendGame, Self::Err>;
 
@@ -1859,7 +1859,7 @@ where
 
     fn send_game(
         &self,
-        chat_id: u32,
+        chat_id: ChatId,
         game_short_name: String,
     ) -> ErasedRequest<'a, SendGame, Self::Err> {
         Requester::send_game(self, chat_id, game_short_name).erase()